fix chunking issues around loadMultipleFromLists & getSharedGroupName
loadMultipleFromLists could be a method on the entity client,
but this way it's easier to test
